+++ frontend/www/scripts/mycoach.js
@@ -1,4 +1,4 @@
-import { sendRequest, getCurrentUser } from "./utils/api.js";
+import { sendRequest, getCurrentUser } from "./utils/requests.js";
 import { HOST } from "./utils/host.js";
 import { createAlert, setReadOnly } from "./utils/dom.js";
