+++ frontend/www/scripts/register.js
@@ -1,4 +1,4 @@
-import { sendRequest } from "./utils/api.js";
+import { sendRequest } from "./utils/requests.js";
 import { HOST } from "./utils/host.js";
 import { createAlert } from "./utils/dom.js";
 import { setCookie } from "./utils/cookies.js";
