+++ frontend/www/scripts/workouts/workouts-page.js
@@ -1,4 +1,4 @@
-import { sendRequest, getCurrentUser } from "../utils/api.js";
+import { sendRequest, getCurrentUser } from "../utils/requests.js";
 import { HOST } from "../utils/host.js";
 
 async function fetchWorkouts(ordering) {
